pyproject.toml
[tool.poetry]
name = "music_bg_extra"
version = "0.2.6"
description = "Extra utils for music_bg"
readme = "README.md"
authors = ["Pavel Kirilin <win10@list.ru>"]
repository = "https://github.com/music-bg/music_bg_extra"
homepage = "https://github.com/music-bg/music_bg_extra"
documentation = "https://github.com/music-bg/music_bg/wiki"
keywords = [
  "Music",
  "Image Processing",
  "Music Background"
]
classifiers = [
  "Topic :: Scientific/Engineering :: Image Processing",
  "Topic :: Multimedia :: Sound/Audio",
  "Typing :: Typed"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Pillow-SIMD = "^7.0.0"
music-bg = "^0.2.0"
numpy = "^1.21.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
mypy = "^0.910"
isort = "^5.9.3"
yesqa = "^1.2.3"
pytest = "^6.2.5"
wemake-python-styleguide = "^0.15.3"
black = "^21.9b0"
pre-commit = "^2.15.0"
autoflake = "^1.4"

[tool.poetry.plugins.mbg_processors]
box_blur = "music_bg_extra.processors.blur:box_blur"
gaussian_blur = "music_bg_extra.processors.blur:gaussian_blur"
circle = "music_bg_extra.processors.circle:circle"
pop_filter = "music_bg_extra.processors.pop:pop_filter"
print = "music_bg_extra.processors.print:img_print"
radial_gradient = "music_bg_extra.processors.gradients:radial_gradient"

[tool.poetry.plugins.mbg_variables]
uuid4 = "music_bg_extra.variables.uuid_gen:uuid4"
most_frequent_color = "music_bg_extra.variables.colors:most_frequent_color_var"
most_frequent_color_inverted = "music_bg_extra.variables.colors:most_frequent_color_inverted_var"
least_frequent_color = "music_bg_extra.variables.colors:most_frequent_color_inverted_var"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
